Extracted routes and added User model method
